ArrayLists for LocalFlatInfo, LocalExpenseGroup and LocalUserProfile replaced with HashMap inside AppData so that corresponding entities can be accessed using Ids.

+++ app/src/main/java/com/example/vivek/rentalmates/interfaces/OnExpenseMembersSelectedReceiver.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface OnExpenseMembersSelectedReceiver {
 
-    public void onOkay(List<Integer> arrayList);
+    public void onOkay(List<Long> arrayList);
 
     public void onCancel();
 }
